various file name fixes, comments and warnings
removed old rendererComponent.cpp(dup) and renamed RendererComponent to SkinRendererComponent
