+++ Editor/src/Ouroboros/Vulkan/SkinRendererSystem.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "RendererComponent.h"
+#include "SkinRendererComponent.h"
 #include "Archetypes_Ecs/src/A_Ecs.h"
 #include "Ouroboros/Scene/Scene.h"
 #include "Ouroboros/Transform/TransformComponent.h"
