+++ Editor/src/Testing/Test/Sample.h
@@ -1,1 +1,14 @@
+/************************************************************************************//*!
+\file           Sample.h
+\project        Ouroboros
+\author         Chua Teck Lee, c.tecklee, 390008420 | code contribution (100%)
+\par            email: c.tecklee\@digipen.edu
+\date           Sept 30, 2022
+\brief          Empty file just to test how files show up when attempting to include
+
+Copyright (C) 2022 DigiPen Institute of Technology.
+Reproduction or disclosure of this file or its contents
+without the prior written consent of DigiPen Institute of
+Technology is prohibited.
+*//*************************************************************************************/
 #pragma once
